requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.10.0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
emoji==1.4.2
et-xmlfile==1.1.0
fpdf==1.7.2
idna==3.2
lxml==4.6.3
numpy==1.21.2
openpyxl==3.0.9
pandas==1.3.3
pyTelegramBotAPI==4.0.0
python-dateutil==2.8.2
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
tornado==6.1
tzdata==2021.1
tzlocal==3.0
urllib3==1.26.6
xlrd==2.0.1
XlsxWriter==3.0.1
emoji